Split tsconfig.json under dom package, prepare for strict mode change
