+++ packages/roosterjs-editor-dom/lib/blockElements/tsconfig.child.json
@@ -1,0 +1,12 @@
+{
+    "compilerOptions": {
+        "strict": false
+    },
+    "extends": "../../../tsconfig.json",
+    "include": ["./**/*.ts"],
+    "references": [
+        { "path": "../../../roosterjs-editor-types/tsconfig.child.json" },
+        { "path": "../selection/tsconfig.child.json" },
+        { "path": "../utils/tsconfig.child.json" }
+    ]
+}
