+++ packages/roosterjs-editor-dom/lib/contentTraverser/tsconfig.child.json
@@ -1,0 +1,14 @@
+{
+    "compilerOptions": {
+        "strict": false
+    },
+    "extends": "../../../tsconfig.json",
+    "include": ["./**/*.ts"],
+    "references": [
+        { "path": "../../../roosterjs-editor-types/tsconfig.child.json" },
+        { "path": "../blockElements/tsconfig.child.json" },
+        { "path": "../inlineElements/tsconfig.child.json" },
+        { "path": "../selection/tsconfig.child.json" },
+        { "path": "../utils/tsconfig.child.json" }
+    ]
+}
