+++ packages/roosterjs-editor-dom/lib/edit/tsconfig.child.json
@@ -1,0 +1,17 @@
+{
+    "compilerOptions": {
+        "strict": false
+    },
+    "extends": "../../../tsconfig.json",
+    "include": ["./**/*.ts"],
+    "references": [
+        { "path": "../../../roosterjs-editor-types/tsconfig.child.json" },
+        { "path": "../selection/tsconfig.child.json" },
+        { "path": "../utils/tsconfig.child.json" },
+        { "path": "../blockElements/tsconfig.child.json" },
+        { "path": "../inlineElements/tsconfig.child.json" },
+        { "path": "../contentTraverser/tsconfig.child.json" },
+        { "path": "../table/tsconfig.child.json" },
+        { "path": "../region/tsconfig.child.json" }
+    ]
+}
