+++ packages/roosterjs-editor-dom/lib/entity/tsconfig.child.json
@@ -1,0 +1,8 @@
+{
+    "compilerOptions": {
+        "strict": false
+    },
+    "extends": "../../../tsconfig.json",
+    "include": ["./**/*.ts"],
+    "references": [{ "path": "../../../roosterjs-editor-types/tsconfig.child.json" }]
+}
