+++ packages/roosterjs-editor-dom/lib/htmlSanitizer/tsconfig.child.json
@@ -1,0 +1,12 @@
+{
+    "compilerOptions": {
+        "strict": false
+    },
+    "extends": "../../../tsconfig.json",
+    "include": ["./**/*.ts"],
+    "references": [
+        { "path": "../../../roosterjs-editor-types/tsconfig.child.json" },
+        { "path": "../utils/tsconfig.child.json" },
+        { "path": "../style/tsconfig.child.json" }
+    ]
+}
