+++ packages/roosterjs-editor-dom/lib/inlineElements/tsconfig.child.json
@@ -1,0 +1,13 @@
+{
+    "compilerOptions": {
+        "strict": false
+    },
+    "extends": "../../../tsconfig.json",
+    "include": ["./**/*.ts"],
+    "references": [
+        { "path": "../../../roosterjs-editor-types/tsconfig.child.json" },
+        { "path": "../utils/tsconfig.child.json" },
+        { "path": "../selection/tsconfig.child.json" },
+        { "path": "../blockElements/tsconfig.child.json" }
+    ]
+}
