+++ packages/roosterjs-editor-dom/lib/list/tsconfig.child.json
@@ -1,0 +1,15 @@
+{
+    "compilerOptions": {
+        "strict": false
+    },
+    "extends": "../../../tsconfig.json",
+    "include": ["./**/*.ts"],
+    "references": [
+        { "path": "../../../roosterjs-editor-types/tsconfig.child.json" },
+        { "path": "../utils/tsconfig.child.json" },
+        { "path": "../region/tsconfig.child.json" },
+        { "path": "../selection/tsconfig.child.json" },
+        { "path": "../contentTraverser/tsconfig.child.json" },
+        { "path": "../style/tsconfig.child.json" }
+    ]
+}
