+++ packages/roosterjs-editor-dom/lib/region/mergeBlocksInRegion.ts
@@ -3,7 +3,7 @@ import contains from '../utils/contains';
 import getBlockElementAtNode from '../blockElements/getBlockElementAtNode';
 import getPredefinedCssForElement from '../htmlSanitizer/getPredefinedCssForElement';
 import getStyles from '../style/getStyles';
-import isNodeInRegion from '../region/isNodeInRegion';
+import isNodeInRegion from './isNodeInRegion';
 import safeInstanceOf from '../utils/safeInstanceOf';
 import setStyles from '../style/setStyles';
 import { BlockElement, RegionBase } from 'roosterjs-editor-types';
