+++ packages/roosterjs-editor-dom/lib/region/tsconfig.child.json
@@ -1,0 +1,16 @@
+{
+    "compilerOptions": {
+        "strict": false
+    },
+    "extends": "../../../tsconfig.json",
+    "include": ["./**/*.ts"],
+    "references": [
+        { "path": "../../../roosterjs-editor-types/tsconfig.child.json" },
+        { "path": "../utils/tsconfig.child.json" },
+        { "path": "../selection/tsconfig.child.json" },
+        { "path": "../contentTraverser/tsconfig.child.json" },
+        { "path": "../blockElements/tsconfig.child.json" },
+        { "path": "../htmlSanitizer/tsconfig.child.json" },
+        { "path": "../style/tsconfig.child.json" }
+    ]
+}
