+++ packages/roosterjs-editor-dom/lib/table/tsconfig.child.json
@@ -1,0 +1,11 @@
+{
+    "compilerOptions": {
+        "strict": false
+    },
+    "extends": "../../../tsconfig.json",
+    "include": ["./**/*.ts"],
+    "references": [
+        { "path": "../../../roosterjs-editor-types/tsconfig.child.json" },
+        { "path": "../utils/tsconfig.child.json" }
+    ]
+}
