+++ packages/roosterjs-editor-dom/lib/utils/wrap.ts
@@ -1,6 +1,6 @@
 import createElement from './createElement';
 import fromHtml from './fromHtml';
-import safeInstanceOf from '../utils/safeInstanceOf';
+import safeInstanceOf from './safeInstanceOf';
 import { CreateElementData, KnownCreateElementDataIndex } from 'roosterjs-editor-types';
 
 /**
